+++ langs/jibberish/jibberishNodes.js
@@ -4,11 +4,11 @@ class additionNode extends jtree.NonTerminalNode {}
 class LineOfCodeNode extends jtree.NonTerminalNode {}
 class SomeNestedNode extends jtree.NonTerminalNode {}
 
-const jibberishNodes = {}
-jibberishNodes.additionNode = additionNode
-jibberishNodes.LineOfCodeNode = LineOfCodeNode
+const nested = {}
+nested.someNestedNode = SomeNestedNode
 
-jibberishNodes.nested = {}
-jibberishNodes.nested.someNestedNode = SomeNestedNode
-
-module.exports = jibberishNodes
+module.exports = {
+  additionNode,
+  LineOfCodeNode,
+  nested
+}