+++ langs/jibberish/jibberishProgram.js
@@ -1,9 +1,0 @@
-const jtree = require("../../index.js")
-
-class jibberishProgram extends jtree.program {
-  executeSync() {
-    return 42
-  }
-}
-
-module.exports = jibberishProgram