+++ langs/jibberish/jibberishProgramRoot.js
@@ -1,0 +1,9 @@
+const jtree = require("../../index.js")
+
+class jibberishProgramRoot extends jtree.programRoot {
+  executeSync() {
+    return 42
+  }
+}
+
+module.exports = jibberishProgramRoot