+++ langs/project/ProjectNodes.js
@@ -1,8 +1,34 @@
 const fs = require("fs")
 const jtree = require("../../index.js")
 const TreeNode = jtree.TreeNode
+const path = require("path")
 
-class projectProgram extends jtree.program {
+class FileNode extends jtree.NonTerminalNode {
+  getFilePath() {
+    return this.cells.filepath.join(" ")
+  }
+
+  _getDependencies() {
+    return this.getChildren()
+      .map(child => {
+        const keyword = child.getKeyword()
+        const childFilePath = child.cells.filepath.join(" ")
+        if (keyword === "external") return ""
+        if (keyword === "absolute") return childFilePath
+        const link = childFilePath
+        const folderPath = jtree.Utils.getPathWithoutFileName(this.getFilePath())
+        const resolvedPath = path.resolve(folderPath + "/" + link)
+        return resolvedPath
+      })
+      .filter(path => path)
+  }
+
+  getMissingDependencies(includedMap) {
+    return this._getDependencies().filter(file => includedMap[file] === undefined)
+  }
+}
+
+class ProjectProgramRoot extends jtree.programRoot {
   getOrderedDependenciesArray() {
     const cloned = this.clone()
     const sorted = []
@@ -73,4 +99,4 @@ class projectProgram extends jtree.program {
   }
 }
 
-module.exports = projectProgram
+module.exports = { ProjectProgramRoot, FileNode }