+++ tests/codemirror.test.js
@@ -4,7 +4,7 @@
 
 const fs = require("fs")
 const GrammarProgram = require("../index.js").getProgramConstructor(__dirname + "/../langs/grammar/grammar.grammar")
-const StampProgram = require("../langs/stamp/index.js")
+const StampProgram = require("../langs/stamp/StampProgram.js")
 const TreeNotationCodeMirrorMode = require("../built/grammar/TreeNotationCodeMirrorMode.js").default
 const TreeUtils = require("../built/base/TreeUtils.js").default